+++ src/http/endpoints.rs
@@ -1,6 +1,5 @@
 use crate::http::client::{ApiClientError, Body, EndpointSchema, EndpointSchemaBuilder, SchemaMethod};
-use crate::transaction::{Currency, SiacoinElement, V1Transaction, V2Transaction};
-use crate::types::{Address, BlockID, Event, Hash256};
+use crate::types::{Address, BlockID, Event, Hash256, Currency, SiacoinElement, V1Transaction, V2Transaction};
 use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
