+++ src/lib.rs
@@ -4,9 +4,6 @@ mod macros;
 pub mod blake2b_internal;
 pub mod encoding;
 pub mod http;
-pub mod specifier;
-pub mod spend_policy;
-pub mod transaction;
 pub mod types;
 
 #[cfg(test)] mod tests;
