+++ src/tests/encoding.rs
@@ -1,7 +1,6 @@
 use crate::blake2b_internal::standard_unlock_hash;
 use crate::encoding::Encoder;
-use crate::spend_policy::{SpendPolicy, UnlockCondition};
-use crate::types::{Address, Hash256, PublicKey};
+use crate::types::{Address, Hash256, PublicKey, SpendPolicy, UnlockCondition};
 use std::convert::TryFrom;
 use std::str::FromStr;
